update:ボスHPバー完成　GameOverHandlerが途中のまま放置されているためテスト不可

+++ Assets/Project/20236027/Script/BossHpFolder/BossHpBarParam.cs
@@ -1,6 +1,6 @@
 public class BossHpBarParam : ParamBase
 {
-    public int bossMaxHp;
-    public int bossNowHp;
+    public float bossMaxHp;
+    public float bossNowHp;
     public string bossName;
 }
